requirements.txt
uvcorn
pyautogen
fastapi
openai
